+++ test/webScriptApi.spec.js
@@ -23,14 +23,17 @@ describe('WebScript', function () {
 
         this.alfrescoJsApi.login('admin', 'admin').then(() => {
             done();
+        }, ()=> {
+
         });
     });
 
     it('execute webScript return 400 error if is not present on the server should be handled by reject promise', function (done) {
         this.webScriptMock.get404Response();
 
-        this.alfrescoJsApi.webScript.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath).then(
-            null,
+        this.alfrescoJsApi.core.webscriptApi.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath).then(()=> {
+            console.log('success');
+        },
             (error) => {
                 expect(error.status).to.be.equal(404);
                 done();
@@ -41,9 +44,12 @@ describe('WebScript', function () {
     it('execute webScript GET return 200 if all is ok  should be handled by resolve promise', function (done) {
         this.webScriptMock.get200Response();
 
-        this.alfrescoJsApi.webScript.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath).then(
+        this.alfrescoJsApi.core.webscriptApi.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath).then(
             () => {
                 done();
+            },
+            (error)=> {
+                console.log('error' + JSON.stringify(error));
             }
         );
     });
@@ -51,42 +57,56 @@ describe('WebScript', function () {
     it('execute webScript that return HTML should not return it as Object', function (done) {
         this.webScriptMock.get200ResponseHTMLFormat();
 
-        this.alfrescoJsApi.webScript.executeWebScript('GET', 'sample/folder/Company%20Home').then((data) => {
-            try {
-                JSON.parse(data);
-            } catch (e) {
-                done();
-            }
-
-        });
+        this.alfrescoJsApi.core.webscriptApi.executeWebScript('GET', 'sample/folder/Company%20Home').then(
+            (data) => {
+                try {
+                    JSON.parse(data);
+                } catch (e) {
+                    done();
+                }
+            }, ()=> {
+                console.log('error');
+            });
     });
 
     describe('Events', function () {
-        it('WebScript should fire done event at the end of an upload', function (done) {
+        it('WebScript should fire success event at the end', function (done) {
             this.webScriptMock.get200Response();
 
-            this.alfrescoJsApi.webScript.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath)
-                .on('success', ()=> {
-                    done();
-                });
+            var webscriptPormise = this.alfrescoJsApi.core.webscriptApi.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath);
+
+            webscriptPormise.catch(()=> {
+            });
+
+            webscriptPormise.on('success', ()=> {
+                done();
+            });
         });
 
         it('WebScript should fire error event if something go wrong', function (done) {
             this.webScriptMock.get404Response();
 
-            this.alfrescoJsApi.webScript.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath)
-                .on('error', ()=> {
-                    done();
-                });
+            var webscriptPormise = this.alfrescoJsApi.core.webscriptApi.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath);
+
+            webscriptPormise.catch(()=> {
+            });
+
+            webscriptPormise.on('error', ()=> {
+                done();
+            });
         });
 
         it('WebScript should fire unauthorized event if get 401', function (done) {
             this.webScriptMock.get401Response();
 
-            this.alfrescoJsApi.webScript.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath)
-                .on('unauthorized', ()=> {
-                    done();
-                });
+            var webscriptPormise = this.alfrescoJsApi.core.webscriptApi.executeWebScript('GET', this.scriptPath, null, this.contextRoot, this.servicePath);
+
+            webscriptPormise.catch(()=> {
+            });
+
+            webscriptPormise.on('unauthorized', ()=> {
+                done();
+            });
         });
     });
 });